made the auto-reconnect functionality slightly more intelligent.  it now won't reconnect if it just reconnected within the last 3 seconds.  so if you're logged in with tnt on emacs process A, and then you log in with tnt (and with the same username) on emacs process B, then the following will happen:
- B connects, A gets disconnected.
- A reconnects, B gets disconnected.
- B reconnects, A gets disconnected.
- A decides not to reconnect again, since it just did a moment before.

so in the end, B is connected and A is not, as should be the case.  it's
too bad that there's no good way to do that with just one login rather
than three, and it is still possible to incur the wrath of the toc server,
winding up unable to log in.  but this is at least an improvement.